Move functions to histogram_utils
The methods _regenerate_histogram and _assimilate_histogram hardly use
self and can be moved into the histogram_utils module.
